requirements.txt
Django
djangorestframework
channels
channels_redis
django-filter
